dmac: support large transfers
In case the transmitted data exceeds the maximum DMA transfer length and
is not a multiple of the latter, we must enqueue DMA transfers of different
length. The data length to receive (given as an argument to `enable_rx()`)
can now exceed the maximum DMA transfer length. The driver transparently
splits this into multiple transfers.