implement Pin driver
The Pin driver for the Zynq-7000 SoC is largely based on the a64_pio_drv
from the genode-allwinner repository. Most of the code could be shared
and went into `common_types.h` and `main.cc`. The platform-specific
parts (register interface, irqs, number of banks and pins) is thereby
kept minimal.